+++ packages/substrate-client/src/chainhead/call.ts
@@ -1,8 +1,9 @@
+import { chainHead } from "@/methods"
 import type { OperationCallDoneRpc } from "./json-rpc-types"
 import { createOperationPromise } from "./operation-promise"
 
 export const createCallFn = createOperationPromise(
-  "chainHead_unstable_call",
+  () => chainHead.call,
   (hash: string, fnName: string, callParameters: string) => [
     [hash, fnName, callParameters],
     (e: OperationCallDoneRpc, res: (output: string) => void) => {